+++ src/frontend/Cargo.toml
@@ -14,6 +14,7 @@ api = { workspace = true }
 async-compat = "0.2"
 async-stream.workspace = true
 async-trait = "0.1"
+auth.workspace = true
 catalog = { workspace = true }
 chrono.workspace = true
 client = { workspace = true }
