+++ src/servers/src/lib.rs
@@ -19,7 +19,6 @@ use datatypes::schema::Schema;
 use query::plan::LogicalPlan;
 use serde::{Deserialize, Serialize};
 
-pub mod auth;
 pub mod configurator;
 pub mod error;
 pub mod grpc;
