+++ src/servers/tests/auth.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use auth::UserInfo;
 use secrecy::ExposeSecret;
 use servers::auth::user_provider::auth_mysql;
 use servers::auth::{
